Fix handling aliases of aliases
For this issue: https://github.com/codesandbox/codesandbox-client/issues/5177

Description of problem:
There are two versions of react-transition-group, one is 2.9.0 and the other is 5.x.x . The 2.9.0 needs dom-helpers@3.4.0 , which we've written down as " react-transition-group needs dom-helpers@3.4.0 ". However, because we have two of react-transition-group, we renamed react-transition-group@2.9.0 to react-transition-group/2.9.0. The problem is that we forgot to rename the alias, so the alias still says react-transition-group needs dom-helpers@3.4.0, but it should be react-transition-group/2.9.0 needs dom-helpers@3.4.0.